requirements.txt
streamlit==1.29.0
numpy==1.26.1
pandas==2.1.2
pillow==10.1.0
nltk==3.8.1
torch>=2.2.0
# Use a pre-built tokenizers wheel
tokenizers>=0.13.3
# Install transformers without tokenizers dependency
transformers>=4.30.0
# Use OpenAI API instead of Replicate
openai>=1.0.0
python-dotenv>=1.0.0 